Fix return type

+++ src/types/Helper/IntegerHelper.php
@@ -25,9 +25,9 @@ class IntegerHelper {
      * Factorial.
      *
      * @param int|null $n The number.
-     * @return int|null Returns the factorial.
+     * @return float|null Returns the factorial.
      */
-    public static function factorial(?int $n): ?int {
+    public static function factorial(?int $n): ?float {
 
         if (null === $n || $n < 0) {
             return null;
